Removed infrastructure code
Added basic template code for init, main and unittest
