+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/__init__.py
@@ -16,3 +16,5 @@
 """
     __init__.py file containing the defaults
 """
+
+from {{cookiecutter.python_import_name}} import main
