+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/__main__.py
@@ -1,0 +1,3 @@
+from {{cookiecutter.python_import_name}} import main
+
+main()
