--- {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/adapter/database/db_connection.py
@@ -1,14 +1,0 @@
-from typing import Any
-
-# from peewee import *
-
-
-class DbConnection:
-    db_connection: Any
-
-    def __init__(self, db_connection: Any) -> None:
-        self.db_connection = db_connection
-        pass
-
-    def get(self) -> Any:
-        return self.db_connection
