+++ {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/app_logging.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2023. Deltares & TNO
+#  Copyright (c) 2023. {cookiecutter.cookiecutter.maintainer_name}}
 #
 #  This program is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
