--- {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/application/configuration.py
@@ -1,11 +1,0 @@
-import os
-from dotenv import dotenv_values
-
-
-def read_environment_config() -> dict:
-    config = {
-        **dotenv_values(".env.shared"),  # load shared development variables
-        **dotenv_values(".env.secret"),  # load sensitive variables
-        **os.environ,  # override loaded values with environment variables
-    }
-    return config
