--- {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/domain/base_configuration_entity.py
@@ -1,3 +1,0 @@
-class BaseConfigurationEntity:
-    def __init__(self) -> None:
-        pass
