--- {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/domain/heat_network.py
@@ -1,9 +1,0 @@
-# import pandapipes
-import logging
-
-LOGGER = logging.getLogger(__name__)
-
-
-class HeatNetwork:
-    def __init__(self) -> None:
-        LOGGER.info("initialized heat network?")
