--- {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/domain/simulation_configuration.py
@@ -1,11 +1,0 @@
-from {{cookiecutter.python_import_name}}.domain.base_configuration_entity import BaseConfigurationEntity
-from dataclasses import dataclass
-import uuid
-
-
-@dataclass
-class SimulationConfiguration(BaseConfigurationEntity):
-    simulation_id: uuid.UUID
-    name: str
-    timestep: float
-    duration: float
