--- {{cookiecutter.repo_name}}/src/{{cookiecutter.python_import_name}}/main.py
@@ -1,5 +1,0 @@
-def main():
-    print('Hello world!')
-
-if __name__ == '__main__':
-    main()
